mnma-index/requirements.txt
langchain
langchain-core
langfuse
langchain-qdrant
langchain_aws==0.1.18
langchain_community==0.2.17
boto3
pydash
pytest-cov
pytest
asyncio==3.4.3
fastapi==0.111.0
fastapi-cognito
qdrant-client
uvicorn[standard]
PyYAML
tenacity
psycopg2-binary
nltk
unstructured[docx,pdf,xlsx]
python-magic
pdf2image
python-dateutil
pytesseract
python-dotenv
amazon-textract-caller>=0.2.0
amazon-textract-textractor
tiktoken
aiofiles>=23.0.0